% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rsadd.R
\name{predict.rsadd}
\alias{predict.rsadd}
\title{Subject-specific prediction from rsadd}
\usage{
\method{predict}{rsadd}(object, newdata, ...)
}
\arguments{
\item{object}{An rsadd object}

\item{newdata}{A data.frame with one row (add covariate values in columns)}

\item{...}{Not used for now}
}
\value{
A data.frame with times, excess and population hazard.
}
\description{
Function
}
\author{
Damjan Manevski \email{damjan.manevski@mf.uni-lj.si}
}
